reduce public footprint; build test
