--- contribs/parking/src/main/java/org/matsim/contrib/parking/parkingproxy/penaltyCalculator/PenaltyFunction.java
@@ -1,30 +1,0 @@
-/* *********************************************************************** *
- * project: org.matsim.*
- * *********************************************************************** *
- *                                                                         *
- * copyright       : (C) 2020 by the members listed in the COPYING,        *
- *                   LICENSE and WARRANTY file.                            *
- * email           : info at matsim dot org                                *
- *                                                                         *
- * *********************************************************************** *
- *                                                                         *
- *   This program is free software; you can redistribute it and/or modify  *
- *   it under the terms of the GNU General Public License as published by  *
- *   the Free Software Foundation; either version 2 of the License, or     *
- *   (at your option) any later version.                                   *
- *   See also COPYING, LICENSE and WARRANTY file                           *
- *                                                                         *
- * *********************************************************************** */
-
-package org.matsim.contrib.parking.parkingproxy.penaltyCalculator;
-
-/**
- * Interface to calculate the penalty a car driver receives depending of the number of cars already there.
- * 
- * @author tkohl / Senozon
- *
- */
-public interface PenaltyFunction {
-	
-	public double calculatePenalty(int numberOfCars);
-}
